requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.1=cpu_0
asgiref=3.3.4=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bigfloat=0.4.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
colorama=0.4.4=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.22=pypi_0
decorator=5.0.5=pyhd3eb1b0_0
django=3.2=pypi_0
facenet-pytorch=2.5.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
gdown=3.12.2=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=2.10=pypi_0
image=1.5.33=pypi_0
intel-openmp=2019.4=243
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jpeg=9b=h024ee3a_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.0=h27cfd23_0
libmklml=2019.0.5=0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
matplotlib=3.4.0=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
online-triplet-loss=0.0.6=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.3=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.7.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py37h2531618_1
readline=8.1=h27cfd23_0
setuptools=52.0.0=py37h06a4308_0
six=1.15.0=py37h06a4308_0
sqlite=3.35.4=hdfb4753_0
sqlparse=0.4.1=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.7.2=py37
torchvision=0.3.0=pypi_0
tornado=6.1=py37h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
